Add folder structure for integrals
