+++ include/Basis/CMakeLists.txt
@@ -1,25 +1,11 @@
 target_sources(gqcp
         PRIVATE
             AOBasis.hpp
-            BaseOneElectronIntegralBuffer.hpp
-            BaseOneElectronIntegralEngine.hpp
-            BaseTwoElectronIntegralBuffer.hpp
-            BaseTwoElectronIntegralEngine.hpp
             CartesianExponents.hpp
             CartesianGTO.hpp
             GTOBasisSet.hpp
             GTOShell.hpp
-            IntegralCalculator.hpp
-            IntegralEngine.hpp
-            LibcintInterfacer.hpp
-            LibcintOneElectronIntegralBuffer.hpp
-            LibcintOneElectronIntegralEngine.hpp
-            LibcintTwoElectronIntegralBuffer.hpp
-            LibcintTwoElectronIntegralEngine.hpp
-            LibintInterfacer.hpp
-            LibintOneElectronIntegralBuffer.hpp
-            LibintOneElectronIntegralEngine.hpp
-            LibintTwoElectronIntegralBuffer.hpp
-            LibintTwoElectronIntegralEngine.hpp
             ShellSet.hpp
         )
+
+add_subdirectory(Integrals)
