+++ include/Basis/Integrals/CMakeLists.txt
@@ -1,0 +1,11 @@
+target_sources(gqcp
+        PRIVATE
+            BaseOneElectronIntegralBuffer.hpp
+            BaseOneElectronIntegralEngine.hpp
+            BaseTwoElectronIntegralBuffer.hpp
+            BaseTwoElectronIntegralEngine.hpp
+            IntegralCalculator.hpp
+            IntegralEngine.hpp
+        )
+
+add_subdirectory(Interfaces)
