+++ src/Basis/LibcintInterfacer.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "Basis/LibcintInterfacer.hpp"
+#include "Basis/Integrals/Interfaces/LibcintInterfacer.hpp"
